+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.3.2
+
+- isEqualsDeep() ; parseFromInlineList() ;
+- parseString() ; parseStringFromInlineList() ;
+- parseBool() ; parseNumsFromInlineList() ; parseIntsFromInlineList() ; parseBoolsFromInlineList() ;
+
 ## 2.3.1
 
 - Clean code.